pyproject.toml
[project]
name = "adaptive-kv"
version = "0.1.0"
description = "Hybrid layer-aware KV cache compression research toolkit"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Adaptive KV Authors" }
]
license = { text = "MIT" }
keywords = ["llm", "kv-cache", "compression", "mlsys"]

dependencies = [
  "numpy>=1.26",
  "torch>=2.3",
  "einops>=0.7",
  "pydantic>=2.6",
  "pyyaml>=6.0",
  "rich>=13.7",
  "tabulate>=0.9",
  "tqdm>=4.66"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-cov>=4.1",
  "ruff>=0.3",
  "mypy>=1.9"
]

[build-system]
requires = ["setuptools>=62", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
pythonpath = ["src"]

